shadps4 : generator : allow .ps4 directory extension
Currently, we only allow to use /userdata/roms/CU12345/gamename.ps4 file.

They are no specific reason that the main directory couldn't be directly the gamename,
shadps4 is able to scan them and find the gameid without problem

This patch add support for

/userdata/roms/gamename.ps4   directory

This is a lot easier to manage game collections, and this is the same behaviour
than for ps3 games.

Also retrobat use this directory extension, so it's easier to share games between retrobat && batocera